Move coucndb repositories to the appropriate classes
