+++ src/main/java/org/opensrp/repository/couch/AllDHIS2Marker.java
@@ -1,4 +1,4 @@
-package org.opensrp.repository;
+package org.opensrp.repository.couch;
 
 import java.util.List;
